move gui_components.h
to components directory
