+++ src/menu/components/file_list.c
@@ -1,6 +1,6 @@
 #include <stdlib.h>
 
-#include "../gui_components.h"
+#include "gui_components.h"
 #include "../fonts.h"
 #include "gui_constants.h"
